Manage slot manipulation centrally and special case replace operations
This is an attempt at fixing https://github.com/w3c/webcomponents/issues/764.

I first wrote an algorithm that combined the operations done by remove and insert. I then used that for replace and replace all. I then abstracted it to avoid duplication.

I'm a little worried that this is not correct as this delays assigning slotables to a slot quite a bit, which I think might be observable in some cases.